add deletion logic
