--- .landscaper/container/pkg/gardenlogin/operation_delete.go
@@ -1,15 +1,0 @@
-// Copyright (c) 2021 SAP SE or an SAP affiliate company. All rights reserved. This file is licensed under the Apache Software License, v. 2 except as noted otherwise in the LICENSE file
-// SPDX-FileCopyrightText: 2021 SAP SE or an SAP affiliate company and Gardener contributors
-//
-// SPDX-License-Identifier: Apache-2.0
-
-package gardenlogin
-
-import (
-	"context"
-)
-
-// Delete runs the delete operation.
-func (o *operation) Delete(ctx context.Context) error {
-	return nil
-}
